Code for Actual and Indexed Image
The defined matlab function for indexing an image is rgb2ind(RGB,n) where a RGB image is converted to a indexed image with n colors. The number of colors should not be more than 65536 and the color map is saved in the map. 
The two functions imagesc(index) and colormap(map) are used to display the indexed image with pixel array and color map respectively. From input to output, RGB has changed to index and n has changed to map.