Code with a Brightness Profile for GrayScale Image
The matlab function to be used for converting a original image to a gray scaled image is rgb2gray(i). Here a three dimensional RGB image is converted to a two dimensional array gray scale image. 
When visualizing the brightness value of the gray scale image in 3D, the brightness depends on the intensity of the pixels. Higher intensity value areas have more brightness than the areas with lower brightness values.