Code for each colour, RGB and GrayScale
An histogram of an image is a kind of statistical graph that can conveniently determine the representation of colours in an image by means of organised colours on x-axis and the amount of the pixels for each colour on y- axis.
JPEG pictures are real colour pictures with 24-bit colour resolution. The image is read in three colour layer matrices: R-, G- and B-color layer when a JPEG format file is read from MATLAB setting using the imread() function. 
In fact the colours R, G and B- are 8-bit; furthermore, the colour channels R, G and B are transformed back into the colour channel 24-bit, utilising the remaining colour channels as black pixel, in order to display the picture in various colour channels.