Code-Original, Equalised and StretchContrast Image
The intensity in pixels is captured by a grayscale image. In the digital processing of images, intensity values range from 0 as the least or darkest intensity, to 255 as the brightest or the largest intensity. In comparison to RGB images with various colour differences only one distinctive aspect within pixels for grayscale images is the difference between dark and light or contrast. 
It is therefore necessary to have a good contrast to capture relevant data and to visualise features better. The differentiation between the highest and the lowest image intensity is described as the contrast. You can calculate this easily from the histogram.