Code for finding Threshold value
Thresholding is the easy way to break the image as well as the most common means of transforming a grey image into an image in binary form. 
In the threshold process we take a threshold value and then we classify all the grey level below that threshold value as 0 (black) and then all the grey level which is equivalent or higher than the threshold value as 1 (white).