Code for Analysing Chocolate Beans
The given image of chocolate beans is loaded and processed. The first task is to apply threshold the given image to emphasis the dark areas in it. First the image is changed to a gray scale image by the usual pre-defined matlab function.

The Thresholds used are 80, 95, 145 and thus the threshold 80 shows a distorted image and also the threshold 145 shows a distorted image. The threshold 95 shows almost a perfect image compared to the original image.