Code for Rectangle, Square, & Circle Shapes
Objects like circles, rectangles and squares should be recognised from the input picture. 
This distinction is determined by the characteristics of increasing element, such as roundness, dimension ratio, centroid etc. We recognise only forms such as circles, rectangles, and squares in this classification from the input image. 
We will therefore focus on the measures we are taking to detect these forms from any input file.