Code for Live Video Processing
We need to learn specific analysis to identify the green colour in each frame. 
The simplest method of achieving so is to transform the whole RGB frame into a specific hypertension plane and remove the RED, Green and Blue pixel values. 
Then select a range that contains various shades of this colour. Nearly all recognisable colours may be identified in a picture in this way.